+++ iOSReferenceApp/ViewModels/Home/ProgramViewModel.swift
@@ -44,6 +44,13 @@ extension ProgramViewModel: LocalizedEntity {
         else if let assetId = program.asset?.assetId { return assetId }
         return "NO TITIE"
     }
+    
+    func anyDescription(locale: String) -> String {
+        if let description = localizedData(locale: locale)?.allDescriptions().last {
+            return description
+        }
+        return localizations().flatMap{ $0.allDescriptions() }.last ?? ""
+    }
 }
 
 extension ProgramViewModel {