+++ CHANGELOG.md
@@ -20,6 +20,7 @@ NO RELEASE DATE SET
 * Listen to onScrubbing event.
 * Cleanup Constants naming (#9)
 * `EMP-10478` Preparation of `DownloadTask`s now occur once `resume()` is called.
+* `EMP-10481` Asset descriptions now fallbacks on any description locale.
 
 ## 0.2.0
 Released 5 Sep 2017